refactor(workspace): extract useWorkspaceSocket hook
Move WebSocket handler, connection lifecycle, and disconnected polling into a hook.
