+++ frontend/src/pages/Workspace/index.tsx
@@ -45,25 +45,16 @@ import DocumentUpload from '@/components/DocumentUpload/DocumentUpload';
 import { ViewModeToggle } from '@/components/ViewMode/ViewModeToggle';
 import MissingDocumentsSection from '@/components/SchemaEditor/MissingDocumentsSection';
 import TableFeedbackWidget from '@/components/TableFeedbackWidget/TableFeedbackWidget';
-import {
-  WS_DISCONNECTED_REFRESH_INTERVAL,
-} from '@/constants';
 import api, { configAPI, loadAPI, schematiqAPI, unitsAPI } from '@/services/api';
-import webSocketService from '@/services/websocket';
 import type {
   CostEstimate,
   DataRow,
   DocumentAvailabilityResponse,
   PaginatedData,
-  ReextractionCompletedData,
-  ReextractionFailedData,
-  ReextractionProgressData,
-  ReextractionStartedData,
   ScheMatiQConfig,
   ScheMatiQStatus,
   SchemaData,
   VisualizationSession,
-  WebSocketMessage,
 } from '@/types';
 import type { DocumentListResponse } from '@/types/unit';
 
@@ -85,6 +76,7 @@ import { SpreadsheetChrome } from './SpreadsheetChrome';
 import { useAddDocuments } from './hooks/useAddDocuments';
 import { useReextraction } from './hooks/useReextraction';
 import { useWorkspaceLayout } from './hooks/useWorkspaceLayout';
+import { useWorkspaceSocket } from './hooks/useWorkspaceSocket';
 import { SpreadsheetSurface } from './SpreadsheetSurface';
 import type {
   CellFormatMap,
@@ -181,7 +173,6 @@ function Workspace() {
     isDraggingDivider,
     startDividerDrag,
   } = useWorkspaceLayout();
-  const [wsConnected, setWsConnected] = useState(false);
 
   const deferredDataRef = useRef<PaginatedData | null>(null);
   const cancelChatPendingRef = useRef<(() => Promise<boolean>) | null>(null);
@@ -357,6 +348,15 @@ function Workspace() {
     toast,
   });
 
+  useWorkspaceSocket({
+    sessionId,
+    refresh,
+    refreshSilent,
+    setActiveSheet,
+    setReextraction,
+    toast,
+  });
+
   useEffect(() => {
     setProjectDialogOpen(!sessionId);
     setSessionMode(requestedMode);
@@ -381,7 +381,6 @@ function Workspace() {
   }, []);
 
   useEffect(() => {
-    setWsConnected(false);
     // Reset session-scoped view/data so a freshly imported/loaded session
     // never renders the previous session's sheet, schema, or rows.
     setActiveSheet('data');
@@ -489,115 +488,6 @@ function Workspace() {
   useEffect(() => {
     refresh();
   }, [refresh, sessionId]);
-
-  useEffect(() => {
-    if (!sessionId || wsConnected) return undefined;
-    const interval = window.setInterval(() => refreshSilent(), WS_DISCONNECTED_REFRESH_INTERVAL);
-    return () => window.clearInterval(interval);
-  }, [refreshSilent, sessionId, wsConnected]);
-
-  useEffect(() => {
-    if (!sessionId) return undefined;
-
-    const handler = (message: WebSocketMessage) => {
-      if (message.type === 'connected') {
-        setWsConnected(true);
-        void refreshSilent();
-        return;
-      }
-
-      if (message.type === 'disconnected' || message.type === 'reconnecting') {
-        setWsConnected(false);
-        return;
-      }
-
-      if (message.type === 'heartbeat' || message.type === 'pong') {
-        return;
-      }
-
-      if (message.type === 'reextraction_started' && message.data) {
-        const payload = message.data as ReextractionStartedData;
-        setReextraction({
-          operationId: payload.operation_id,
-          columns: payload.columns || [],
-          progress: 0,
-          processedDocuments: 0,
-          totalDocuments: payload.total_documents || 0,
-        });
-        setActiveSheet('data');
-        void refresh({ silent: true });
-        return;
-      }
-
-      if (message.type === 'reextraction_progress' && message.data) {
-        const payload = message.data as ReextractionProgressData;
-        setReextraction((current) => ({
-          operationId: payload.operation_id,
-          columns: current?.columns || (payload.column ? [payload.column] : []),
-          progress: payload.progress ?? current?.progress ?? 0,
-          processedDocuments: payload.processed_documents ?? current?.processedDocuments ?? 0,
-          totalDocuments: payload.total_documents ?? current?.totalDocuments ?? 0,
-          currentColumn: payload.column || current?.currentColumn,
-        }));
-        void refresh({ silent: true });
-        return;
-      }
-
-      if (message.type === 'reextraction_completed' && message.data) {
-        const payload = message.data as ReextractionCompletedData;
-        setReextraction(null);
-        void refresh({ silent: true });
-        toast({
-          title: 'Re-extraction completed',
-          description: payload.columns?.length
-            ? `Updated ${payload.columns.length} column(s) from source documents.`
-            : 'Table values were refreshed from source documents.',
-        });
-        return;
-      }
-
-      if (message.type === 'reextraction_failed' && message.data) {
-        const payload = message.data as ReextractionFailedData;
-        setReextraction(null);
-        toast({
-          title: 'Re-extraction failed',
-          description: payload.error || 'Could not re-extract values from source documents.',
-          variant: 'destructive',
-        });
-        return;
-      }
-
-      if (message.type === 'reextraction_stopped') {
-        setReextraction(null);
-        void refresh({ silent: true });
-        return;
-      }
-
-      if (
-        message.type === 'progress' ||
-        message.type === 'completed' ||
-        message.type === 'schema_completed' ||
-        message.type === 'schema_updated' ||
-        message.type === 'cell_extracted' ||
-        message.type === 'row_completed' ||
-        message.type === 'schema_progress' ||
-        message.type === 'reprocessing_progress' ||
-        message.type === 'reprocessing_completed' ||
-        message.type === 'observation_unit_definition_updated'
-      ) {
-        void refresh({ silent: true });
-      }
-    };
-
-    webSocketService.addMessageHandler(handler);
-    webSocketService.connect(sessionId, 'progress');
-
-    return () => {
-      webSocketService.removeMessageHandler(handler);
-      webSocketService.disconnect();
-      setWsConnected(false);
-    };
-  }, [refresh, refreshSilent, sessionId, toast]);
 
   const estimateCurrentCost = useCallback(async () => {
     if (!sessionId) return;
